Move shared Zod helpers and remove unused test helpers
