+++ src/lazy.ts
@@ -6,12 +6,15 @@ import { isNamedType } from './graphql.js';
 
 /**
  * Wraps a schema expression in a library-specific lazy reference when the type
- * is a complex (non-scalar, non-enum) named type — avoiding issues with
+ * is a complex (non-scalar, non-enum) named type, avoiding issues with
  * mutually-recursive input types.
  *
  * Each validation library has its own lazy syntax (z.lazy, v.lazy, etc.), so
  * callers supply the wrapper function.
  *
+ * @param visitor - Type lookup used to distinguish scalars/enums from complex types.
+ * @param type - GraphQL type node for the schema expression.
+ * @param schema - Generated schema expression to wrap when needed.
  * @param lazyWrapper - e.g. `(s) => \`z.lazy(() => ${s})\``
  */
 export function buildMaybeLazy(
